backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# Doküman işleme
PyMuPDF==1.24.10
python-docx==1.1.2

# LLM & Embedding
groq==0.11.0
google-generativeai==0.8.3

# Vektör DB & RAG
chromadb==0.5.15
langchain==0.3.3
langchain-community==0.3.2
langchain-google-genai==2.0.1
langchain-groq==0.2.1

# Yardımcılar
python-dotenv==1.0.1
httpx==0.27.2
